Refactor data model
Now the models are separate files and the old controller has been deleted.

The recipe and steps types are now forced by the database and can be changed only by changing or adding rows into the database. Also started adding support to localisation.

Updated validation to the new constraints.

The inputs on the form templates that correspond to arrays now have "[]" in the name so arrays are forced even with one element.