Fix little bugs
When HTTPS is not enabled, did not fall back to NOT using cache and did not work.

Added minum width to the ingredients as it was displayed bad in some devices.

Add forgotten HTTPS_PORT to README

Fix styles in edit form because the css was linked in the wrong order.

Fix conversion error. When an ingredient did not have a value/unit the
converstion set the value to 0 when displaying the original value.